+++ src/main/java/dev/galacticraft/mod/network/c2s/BubbleMaxPayload.java
@@ -35,7 +35,7 @@ import net.minecraft.resources.ResourceLocation;
 import org.jetbrains.annotations.NotNull;
 
 public record BubbleMaxPayload(int max) implements C2SPayload {
-    public static final StreamCodec<ByteBuf, BubbleMaxPayload> STREAM_CODEC = ByteBufCodecs.INT.map(BubbleMaxPayload::new, packet -> packet.max);
+    public static final StreamCodec<ByteBuf, BubbleMaxPayload> STREAM_CODEC = ByteBufCodecs.BYTE.map(BubbleMaxPayload::new, packet -> (byte)packet.max);
     public static final ResourceLocation ID = Constant.id("bubble_max");
     public static final CustomPacketPayload.Type<BubbleMaxPayload> TYPE = new CustomPacketPayload.Type<>(ID);
